Climate

Streamfall.climate_values — Method

climate_values(node::NetworkNode, climate::Climate, timestep::Int)

Extract climate related data for a given time step.

Streamfall.climate_values — Method

climate_values(node::NetworkNode, climate::Climate)

Extract climate related data.

Streamfall.extract_climate — Method

Create a climate dataset of Precipitation (P) and Potential Evapotranspiration (PET). Data for multiple gauges may be defined in a single dataset.

P and PET columns are identified by _P and _PET suffixes by default.

Arguments

data : Observation data

P_suffix : Suffix used to indicate precipitation (default: "_P")

PET_suffix : Suffix used to indicate Potential Evapotranspiration (default: "_PET")

T_suffix : Suffix used to indicate Temperature (default: "_PET")

Returns

Climate

Streamfall.extract_flow — Function

Extract streamflow data from file.

Streamflow (Q) column is identified the Gauge ID.

Flow data is identified with the suffix _Q by default e.g., ("000001_Q")

Arguments

data : Observation data

gauge_id : Gauge/Node ID

suffix : Suffix used to indicate flow data (default: "_Q")

Returns

DataFrame of observations for selected gauge.

Streamfall.rainfall_data — Method

rainfall_data(node::NetworkNode, climate::Climate)::DataFrame

Extract rainfall data for a given node.

Streamfall.sim_length — Method

sim_length(climate::Climate)::Int64

Simulation length is dependent on available climate data.

Streamfall.subcatchment_data — Method

subcatchment_data(node::NetworkNode, climate::Climate)

Extract all data for a given node from climate object.