IHACRES

Streamfall.extract_spec! — Method

extract_spec!(node::DamNode, spec::AbstractDict)::Nothing

Additional processing to extract IHACRES-specific details.

Streamfall.param_info — Method

param_info(node::IHACRESNode)::Tuple

Extract node parameter names, values, and bounds for IHACRESNode types.

Streamfall.reset! — Method

reset!(s_node::IHACRESNode)::Nothing

Reset node. Clears all states back to their initial values.

Streamfall.run_node_with_temp! — Function

run_node_with_temp!(
    s_node::IHACRESBilinearNode,
    rain::Float64,
    temp::Float64,
    inflow::Float64,
    ext::Float64,
    gw_exchange::Float64=0.0;
    current_store=nothing,
    quick_store=nothing,
    slow_store=nothing,
    gw_store=nothing
)::Tuple{Float64,Float64}

Run node with temperature data to calculate outflow and update state.

Streamfall.run_timestep! — Method

run_timestep!(
    node::IHACRESNode, climate::Climate, timestep::Int64;
    inflow=nothing, extraction=nothing, exchange=nothing
)::Float64

Run the given IHACRESBilinearNode for a timestep.

Streamfall.update_params! — Method

update_params!(
    node::IHACRESBilinearNode,
    d::F,
    d2::F,
    e::F,
    f::F,
    a::F,
    b::F,
    s_coef::F,
    alpha::F
)::Nothing where {F<:Float64}

Update model parameters.

Streamfall.IHACRESBilinearNode — Method

IHACRESBilinearNode(name::String, area::Float64, d::Float64, d2::Float64, e::Float64, f::Float64,
            a::Float64, b::Float64, s_coef::Float64, alpha::Float64,
            store::Float64, quick::Float64, slow::Float64, gw_store::Float64)

Create a IHACRES node that adopts the bilinear CMD module.