Plotting

Streamfall.plot_residuals — Method

plot_residuals(obs::Array, sim::Array; xlabel="", ylabel="", title="")

Plot residual between two sequences.

Arguments

x : x-axis data

y : y-axis data

xlabel : x-axis label

ylabel : y-axis label

title : title text

Streamfall.symlog — Method

Symmetrical log values.

https://kar.kent.ac.uk/32810/2/2012Bi-symmetric-log-transformationv5.pdf https://discourse.julialang.org/t/symmetrical-log-plot/45709/3

Streamfall.temporal_cross_section — Method

temporal_cross_section(
    dates, obs, sim;
    title="", ylabel="Median Error", label=nothing, period::Function=monthday, kwargs...
)

Provides indication of temporal variation and uncertainty across time, grouped by period.

Notes: Assumes daily data. Filters out leap days.

Arguments

dates : Date of each observation

obs : Observed data

sim : Simulated data

title : Optional plot title. Blank if not provided.

ylabel : Optional replacement ylabel. Uses name of period if not provided.

label : Optional legend label. Uses ylabel if not provided.

period : Method from Dates package to group (defaults to month)

Streamfall.temporal_cross_section — Method

temporal_cross_section(dates, obs; ylabel=nothing, period::Function=monthday)

Provides indication of temporal variation and uncertainty across time, grouped by period.

Notes: Assumes daily data. Filters out leap days.

Arguments

dates : Date of each observation

obs : observed data

ylabel : Optional replacement ylabel. Uses name of func if not provided.

period::Function : Method from Dates package to group (defaults to monthday)