GR4J

Streamfall.reset! — Method

reset!(node::GR4JNode)::Nothing

Reset node. Clears all states back to their initial values.

Streamfall.run_gr4j — Function

run_gr4j(P::Float64, E::Float64,
         X1::Float64, X2::Float64, X3::Float64, X4::Float64, area::Float64,
         p_store::Float64=0.0, r_store::Float64=0.0)::Tuple

Generated simulated streamflow for given rainfall and potential evaporation.

Parameters

P : Catchment average rainfall

E : Catchment average potential evapotranspiration

X1 - X4 : X parameters

area : Catchment area

p_store : Initial production store

r_store : Initial state store

Returns

tuple of simulated outflow [ML/day], and intermediate states: pstore, rstore, UH1, UH2

Streamfall.run_node! — Method

run_node!(node::GR4JNode, climate::Climate, timestep::Int;
          inflow::Float64, extraction::Float64, exchange::Float64)

Run given GR4J node for a time step.

Streamfall.run_node! — Method

run_node!(node::GR4JNode, climate::Climate)

Run GR4J node for all time steps in given climate sequence.

Streamfall.s_curve — Method

s_curve(t::Float64, x4::Float64, uh2::Bool = false)::Float64

Determine unit hydrograph ordinates.

Streamfall.GR4JNode — Type

GR4J Node

A four-parameter model with two stores.

Parameters

x1 : maximum capacity of the production store (mm) (> 0)

x2 : groundwater exchange coefficient (mm) (value < and > 0 possible)

x3 : one day ahead maximum capacity of the routing store (mm, > 0)

x4 : time base of unit hydrograph UH1 (days, > 0.5)

References

Perrin, C., Michel, C., Andréassian, V., 2003. Improvement of a parsimonious model for streamflow simulation. Journal of Hydrology 279, 275–289. https://doi.org/10.1016/S0022-1694(03)00225-7

MacDonald, A. 2014. Python GR4J https://github.com/amacd31/gr4j