IHACRES

Streamfall.param_info — Method

param_info(node::IHACRESNode)::Tuple

Extract node parameter names, values, and bounds for IHACRESNode types.

Streamfall.reset! — Method

reset!(s_node::IHACRESNode)::Nothing

Reset node. Clears all states back to their initial values.

Streamfall.run_node! — Method

run_node!(
    node::IHACRESNode, climate::Climate, ts::Int;
    inflow=nothing, extraction=nothing, exchange=nothing
)

Run a specific node for a specified time step.

Arguments

node::IHACRESNode :

climate::Climate :

ts::Int : current time step

inflow::DataFrame : Time series of inflows from any upstream node.

extraction::DataFrame : Time series of water orders (expects column of _releases)

exchange::DataFrame : Time series of groundwater flux

Streamfall.run_node_with_temp! — Function

run_node_with_temp!(s_node::IHACRESBilinearNode,
                    rain::Float64,
                    temp::Float64,
                    inflow::Float64,
                    ext::Float64,
                    gw_exchange::Float64=0.0;
                    current_store=nothing,
                    quick_store=nothing,
                    slow_store=nothing
                    gw_store=nothing)::Tuple{Float64, Float64}

Run node with temperature data to calculate outflow and update state.

Streamfall.run_timestep! — Method

run_timestep!(
    node::IHACRESBilinearNode, climate::Climate, timestep::Int,
    inflow::Float64, extraction::Float64, exchange::Float64
)

Run the given IHACRESBilinearNode for a time step.

Streamfall.update_params! — Method

update_params!(node::IHACRESBilinearNode, d::Float64, d2::Float64, e::Float64, f::Float64,
               a::Float64, b::Float64, s_coef::Float64, alpha::Float64)::Nothing

Update model parameters.

Streamfall.IHACRESBilinearNode — Method

IHACRESBilinearNode(name::String, area::Float64, d::Float64, d2::Float64, e::Float64, f::Float64,
            a::Float64, b::Float64, s_coef::Float64, alpha::Float64,
            store::Float64, quick::Float64, slow::Float64, gw_store::Float64)

Create a IHACRES node that adopts the bilinear CMD module.

IHACRES - Expuh

Streamfall.run_node! — Method

run_node!(
    s_node::ExpuhNode,
    rain::F,
    evap::F,
    inflow::F,
    ext::F,
    gw_exchange::F;
    current_store=nothing,
    quick_store=nothing,
    slow_store=nothing,
)::F where {F<:AbstractFloat}

Run given IHACRES ExpuhNode for a time step based on last known state.