Generic node methods

Streamfall.create_node — Method

Create an arbitrary node.

Streamfall.get_node — Method

get_node(sn::StreamfallNetwork, node_name::String)

Retrieve node_id and node property for a specified gauge.

Arguments

sn : Streamfall Network

node_name : name of node of interest

Returns

Tuple, node position id and node object

Streamfall.get_node_id — Method

Retrieve network node_id for a given gauge (by name).

Streamfall.param_info — Method

param_info(node::NetworkNode)

Generic parameter information extractor.

Extracts parameter names, values, and bounds

Streamfall.reset! — Method

reset!(node::SYMHYDNode)::Nothing

Reset node. Clears all states back to their initial values.

Streamfall.run_node! — Method

Run SYMHYD for all time steps within a climate scenario.

Streamfall.run_symhyd — Method

Run SYMHYD for a single time step with given inputs and state variables.

Streamfall.run_timestep! — Method

Run SYMHYD for a given time step

Streamfall.subcatchment_data — Function

subcatchment_data(node::NetworkNode, data::DataFrame)

Extract all data for a given node from a dataframe.

Streamfall.SYMHYDNode — Type

Streamfall.SYMHYDNode — Method

Create node from spec.