Input data format

By convention, Streamfall expects data to be formatted in the following manner:

All time series must have a "Date" column in YYYY-mm-dd format.

By convention, column names follow the pattern: [node_name]_[phenomenon]_[other metadata]

Unit of measure itself is optionally included in square brackets ([])

In-file comments are indicated with a hash (#)

Streamflow data

Data for a node representing a subcatchment or gauge should, at a minimum, include columns for date and streamflow (Q) in a consistent unit (in this case, megaliters/day).

Streamflow columns should begin with its name and include an identifiable suffix. Here, _Q is used but this can be user-defined when preparing data for modelling.

# A comment
Date,406219_Q_[ML]
1988-07-23,0.5184
1988-07-24,13.6512
1988-07-25,407.6352
1988-07-26,2488.4928
1988-07-27,3443.6448
... snip ...

Dam releases

Note that releases is denoted with _releases.

# A comment
Date, 406000_releases_[ML], 406000_extractions_[ML]
1981-01-01,0,0
1981-01-02,359,0
1981-01-03,359,0
1981-01-04,359,0
1981-01-05,460,0
... snip ...

Climate data

Streamfall currently expects data for all nodes to be stored in one large DataFrame.

Example climate data structure

Date, 406214_P, 406214_ET, 406219_P, 406219_ET
1981-01-01, 0.0, 4.8, 0.0, 4.9
1981-01-02, 0.1, 0.5, 0.1, 3.3
1981-01-03, 10.5, 5.3, 7.2, 2.3
1981-01-04, 9.89, 7.9, 6.1, 4.3
1981-01-05, 0.3, 4.2, 0.2, 6.4
... snip ...

where P indicates rainfall, and ET denotes evapotranspiration.

Climate data functions

Streamfall.climate_values — Method

climate_values(node::NetworkNode, climate::Climate, timestep::Int)

Extract climate related data for a given time step.

Streamfall.climate_values — Method

climate_values(node::NetworkNode, climate::Climate)

Extract climate related data.

Streamfall.extract_climate — Method

Create a climate dataset of Precipitation (P) and Potential Evapotranspiration (PET). Data for multiple gauges may be defined in a single dataset.

P and PET columns are identified by _P and _PET suffixes by default.

Arguments

data : Observation data

P_suffix : Suffix used to indicate precipitation (default: "_P")

PET_suffix : Suffix used to indicate Potential Evapotranspiration (default: "_PET")

T_suffix : Suffix used to indicate Temperature (default: "_PET")

Returns

Climate

Streamfall.extract_flow — Function

Extract streamflow data from file.

Streamflow (Q) column is identified the Gauge ID.

Flow data is identified with the suffix _Q by default e.g., ("000001_Q")

Arguments

data : Observation data

gauge_id : Gauge/Node ID

suffix : Suffix used to indicate flow data (default: "_Q")

Returns

DataFrame of observations for selected gauge.

Streamfall.rainfall_data — Method

rainfall_data(node::NetworkNode, climate::Climate)::DataFrame

Extract rainfall data for a given node.

Streamfall.sim_length — Method

sim_length(climate::Climate)::Int64

Simulation length is dependent on available climate data.

Streamfall.subcatchment_data — Method

subcatchment_data(node::NetworkNode, climate::Climate)

Extract all data for a given node from climate object.