Implemented node types

IHACRES

Streamfall.param_info — Method

Extract node parameter values and bounds

Streamfall.run_node! — Function

Run node with ET data to calculate outflow and update state.

Parameters

snode rain: float, rainfall evap: float, evapotranspiration inflow: float, inflow from previous node ext: float, irrigation and other water extractions gwexchange: float, flux in ML - positive is contribution to stream, negative is infiltration currentstore: replacement cmd value quickstore: replacement quickstore value slowstore: replacement slow_store value

Returns

float, outflow from node

Streamfall.run_node_with_temp! — Function

Run node with temperature data to calculate outflow and update state.

Parameters

snode rain: float, rainfall temp: float, temperature inflow: float, inflow from previous node ext: float, irrigation and other water extractions gwexchange: float, flux in ML - positive is contribution to stream, negative is infiltration currentstore: replacement cmd value quickstore: replacement quickstore value slowstore: replacement slow_store value

Returns

float, outflow from node

Streamfall.update_params! — Method

Streamfall.update_params! — Method

Update all parameters

IHACRES - Expuh

Dam Level

Streamfall.param_info — Method

Extract node parameter values and bounds

Streamfall.run_node! — Function

Calculate outflow for the dam node for a single time step.

Parameters

node : DamNode rain : Float64, rainfall in mm et : Float64, evapotranspiration data in mm irrigext : Float64, irrigation extractions extractions : Float64, extraction data in ML gwflux : Float64, groundwater interaction

:returns: numeric, outflow from Dam

Streamfall.update_params! — Method

Streamfall.update_volume — Method

Update dam volume for timestep

Parameters

volume : float, current water volume in ML nodeinflow : float, inflow from previous node in ML gamma : float, groundwater exchange (positive is gain from gw flow, negative is loss to infiltration) rain : float, rainfall input evap : float, evaporation loss infiltration : float, infiltration loss area : float, dam surface area in square kilometers extractions : float, water extraction from dam in ML discharge : float, discharge from dam in ML maxstore : float, maximum dam storage in ML

Returns

float, volume of water stored in dam