IHACRES

Streamfall.param_info — Method

param_info(node::IHACRESNode; with_level::Bool = true)::Tuple

Extract node parameter names, values, and bounds for IHACRESNode types.

Streamfall.reset! — Method

reset!(s_node::IHACRESNode)::Nothing

Reset node. Clears all states back to their initial values.

Streamfall.run_node! — Method

run_node!(node::IHACRESNode, climate::Climate, timestep::Int; 
              inflow=nothing, extraction=nothing, exchange=nothing)

Run a specific node for a specified time step.

Arguments

node::IHACRESNode :

climate::Climate :

timestep::Int : current time step

inflow::DataFrame : Time series of inflows from any upstream node.

extraction::DataFrame : Time series of water orders (expects column of _releases)

exchange::DataFrame : Time series of groundwater flux

Streamfall.run_node_with_temp! — Function

run_node_with_temp!(s_node::BilinearNode,
                    rain::Float64,
                    temp::Float64,
                    inflow::Float64,
                    ext::Float64,
                    gw_exchange::Float64=0.0;
                    current_store=nothing,
                    quick_store=nothing,
                    slow_store=nothing
                    gw_store=nothing)::Tuple{Float64, Float64}

Run node with temperature data to calculate outflow and update state.

Streamfall.run_step! — Method

run_step!(s_node::BilinearNode,
          rain::Float64,
          evap::Float64,
          inflow::Float64,
          ext::Float64,
          gw_exchange::Float64;
          current_store::Union{Nothing, Float64}=nothing,
          quick_store::Union{Nothing, Float64}=nothing,
          slow_store::Union{Nothing, Float64}=nothing,
          gw_store::Union{Nothing, Float64}=nothing)::Tuple{Float64, Float64}

Run node with ET data to calculate outflow and update state.

Arguments

s_node::BilinearNode : IHACRESNode

rain : rainfall for time step

evap : evapotranspiration for time step

inflow : inflow from previous node

ext : irrigation and other water extractions

gw_exchange : flux in ML where positive is contribution to stream, negative is infiltration

current_store : replacement cmd state value (uses last known state if not provided)

quick_store : replacement quick store state value

slow_store : replacement slow store state value

gw_store : replacement groundwater store state value

Returns

float, outflow from node [ML/day], stream level

Streamfall.run_timestep! — Method

run_timestep!(s_node::BilinearNode,
              rain::Float64,
              evap::Float64,
              timestep::Union{Int64, Nothing};
              inflow::Float64,
              extraction::Float64,
              exchange::Float64)::Tuple{Float64, Float64}

Run node for a given time step.

Streamfall.update_params! — Method

update_params!(node::BilinearNode, d::Float64, d2::Float64, e::Float64, f::Float64,
               a::Float64, b::Float64, s_coef::Float64, alpha::Float64)::Nothing

Update model parameters.

Streamfall.update_params! — Method

update_params!(node::BilinearNode{Param}, d::Float64, d2::Float64, e::Float64, f::Float64,
               a::Float64, b::Float64, s_coef::Float64, alpha::Float64,
               p1::Float64, p2::Float64, p3::Float64, p4::Float64, p5::Float64, p6::Float64, p7::Float64, p8::Float64, CTF::Float64)::Nothing

Update all parameters.

Streamfall.BilinearNode — Method

BilinearNode(name::String, area::Float64, d::Float64, d2::Float64, e::Float64, f::Float64, 
            a::Float64, b::Float64, s_coef::Float64, alpha::Float64, 
            store::Float64, quick::Float64, slow::Float64, gw_store::Float64)

Create a IHACRES node that adopts the bilinear CMD module.

IHACRES - Expuh

Streamfall.run_node! — Method

run_node!(s_node::ExpuhNode, rain::Float64, evap::Float64, 
          inflow::Float64, ext::Float64, gw_exchange::Float64;
          current_store=nothing,
          quick_store=nothing,
          slow_store=nothing,
        )::Tuple

Run given IHACRES ExpuhNode for a time step based on last known state.