Included metrics

Streamfall.ADJ_R2 — Method

Determine adjusted R^2

Arguments

obs::Vector : observations

sim::Vector : modeled results

p::Int : number of explanatory variables

Streamfall.BKGE — Method

Bounded KGE, bounded between -1 and 1.

Arguments

obs::Vector : observations

sim::Vector : modeled results

Streamfall.BmKGE — Method

Bounded modified KGE between -1 and 1.

Arguments

obs::Vector : observations

sim::Vector : modeled results

Streamfall.BnpKGE — Method

Bounded non-parametric KGE between -1 and 1.

Arguments

obs::Vector : observations

sim::Vector : modeled results

Streamfall.EV — Method

EV(obs, sim)

Explained Variance.

Represents the amount of variation in the observations which the predictions are able to explain.

Streamfall.KGE — Method

KGE(obs::Vector, sim::Vector; scaling::Tuple=nothing)::Float64

Calculate the 2009 Kling-Gupta Efficiency (KGE) metric.

Decomposes NSE into correlation (r), relative variability (α), and bias (β) terms.

A KGE score of 1 means perfect fit. A score < -0.41 indicates that the mean of observations provides better estimates (see Knoben et al., 2019).

The scaling argument expects a three-valued tuple which scales r, α and β factors respectively. If not specified, defaults to 1.

Note: Although similar, NSE and KGE cannot be directly compared.

References

Gupta, H.V., Kling, H., Yilmaz, K.K., Martinez, G.F., 2009. Decomposition of the mean squared error and NSE performance criteria: Implications for improving hydrological modelling. Journal of Hydrology 377, 80–91. https://doi.org/10.1016/j.jhydrol.2009.08.003

Knoben, W.J.M., Freer, J.E., Woods, R.A., 2019. Technical note: Inherent benchmark or not? Comparing Nash-Sutcliffe and Kling-Gupta efficiency scores (preprint). Catchment hydrology/Modelling approaches. https://doi.org/10.5194/hess-2019-327

Mizukami, N., Rakovec, O., Newman, A.J., Clark, M.P., Wood, A.W., Gupta, H.V., Kumar, R., 2019. On the choice of calibration metrics for “high-flow” estimation using hydrologic models. Hydrology and Earth System Sciences 23, 2601–2614. https://doi.org/10.5194/hess-23-2601-2019

Streamfall.LME — Method

Liu Mean Efficiency metric (LME).

Reformulation of the KGE metric said to be advantageous for capturing extreme flow events.

Arguments

obs::Vector : observations

sim::Vector : modeled results

References

Liu, D., 2020. A rational performance criterion for hydrological model. Journal of Hydrology 590, 125488. https://doi.org/10.1016/j.jhydrol.2020.125488

Streamfall.MAE — Method

Mean Absolute Error

Streamfall.ME — Method

Mean Error

Streamfall.NKGE — Method

Normalized KGE between 0 and 1.

Arguments

obs::Vector : observations

sim::Vector : modeled results

Streamfall.NNSE — Method

Normalized Nash-Sutcliffe Efficiency score (bounded between 0 and 1).

References

Nossent, J., Bauwens, W., 2012. Application of a normalized Nash-Sutcliffe efficiency to improve the accuracy of the Sobol’ sensitivity analysis of a hydrological model. EGU General Assembly Conference Abstracts 237.

Streamfall.NSE — Method

The Nash-Sutcliffe Efficiency score

Streamfall.NmKGE — Method

Normalized modified KGE between 0 and 1.

Arguments

obs::Vector : observations

sim::Vector : modeled results

Streamfall.NnpKGE — Method

Normalized non-parametric KGE between 0 and 1.

Arguments

obs::Vector : observations

sim::Vector : modeled results

Streamfall.PBIAS — Method

PBIAS(obs::Vector, sim::Vector)::Float64

Percent bias between sim and obs

Model performance for streamflow can be determined to be satisfactory if the Nash-Sutcliffe Efficiency (NSE) score > 0.5, the RMSE standard deviation ratio (RSR) < 0.7 and percent bias (PBIAS) is +/- 25% (see [1]).

References

Moriasi, D.N., Arnold, J.G., Liew, M.W.V., Bingner, R.L., Harmel, R.D., Veith, T.L., 2007. Model Evaluation Guidelines for Systematic Quantification of Accuracy in Watershed Simulations. Transactions of the ASABE 50, 885–900. https://doi.org/10.13031/2013.23153

Streamfall.R2 — Method

Coefficient of determination (R^2)

Aliases NSE()

Streamfall.RMSE — Method

Root Mean Square Error

Streamfall.RSR — Method

RSR(obs::Vector, sim::Vector)::Float64

The RMSE-observations standard deviation ratio (RSR).

Varies between 0 and a large positive value, where 0 indicates an RMSE value of 0.

References

Moriasi, D.N., Arnold, J.G., Liew, M.W.V., Bingner, R.L., Harmel, R.D., Veith, T.L., 2007. Model Evaluation Guidelines for Systematic Quantification of Accuracy in Watershed Simulations. Transactions of the ASABE 50, 885–900. https://doi.org/10.13031/2013.23153

Streamfall.inverse_metric — Method

inverse_metric(obs, sim; metric, comb_method::Function=mean)

A meta-objective function which combines the performance of the given metric as applied to the discharge and the inverse of the discharge.

By default, the combination method is to take the mean.

References

Garcia, F., Folton, N., Oudin, L., 2017. Which objective function to calibrate rainfall–runoff models for low-flow index simulations? Hydrological Sciences Journal 62, 1149–1166. https://doi.org/10.1080/02626667.2017.1308511

Streamfall.mKGE — Method

Calculate the modified KGE metric (2012).

Also known as KGE prime (KGE').

Arguments

obs::Vector: observations

sim::Vector : modeled results

scaling::Tuple : scaling factors in order of timing (r), magnitude (β), variability (γ). Defaults to (1,1,1).

References

Kling, H., Fuchs, M., Paulin, M., 2012. Runoff conditions in the upper Danube basin under an ensemble of climate change scenarios. Journal of Hydrology 424–425, 264–277. https://doi.org/10.1016/j.jhydrol.2012.01.011

Streamfall.mean_NmKGE — Method

Mean Inverse NmKGE

Said to produce better fits for low-flow indices compared to mKGE (see [1]).

Arguments

obs::Vector : observations

sim::Vector : modeled results

scaling::Tuple : scaling factors for r, α, and β (defaults to 1.0)

ϵ::Float64 : small constant to use with inverse flow to allow consideration of periods with no flow.

References

Garcia, F., Folton, N., Oudin, L., 2017. Which objective function to calibrate rainfall–runoff models for low-flow index simulations? Hydrological Sciences Journal 62, 1149–1166. https://doi.org/10.1080/02626667.2017.1308511

Streamfall.naive_split_metric — Method

Naive approach to split metrics.

Split metrics are a meta-objective optimization approach which "splits" data into subperiods. The objective function is calculated for each subperiod and then recombined. The approach addresses the lack of consideration of dry years with least-squares.

In Fowler et al., [1] the subperiod is one year. The implementation offered here is "naive" in that the data is partitioned into N chunks of n_members and does not consider date/time.

Arguments

obs::Vector : Historic observations to compare against

sim::Vector : Modeled time series

n_members::Int : number of members per chunk (i.e., sub-samples), defaults to 365

metric::Function : Objective function to apply, defaults to NNSE

comb_method::Function : Recombination method, defaults to mean

References

Fowler, K., Peel, M., Western, A., Zhang, L., 2018. Improved Rainfall-Runoff Calibration for Drying Climate: Choice of Objective Function. Water Resources Research 54, 3392–3408. https://doi.org/10.1029/2017WR022466

Streamfall.npKGE — Method

Calculate the non-parametric Kling-Gupta Efficiency (KGE) metric.

Arguments

obs::Vector : observations

sim::Vector : modeled

scaling::Tuple : scaling factors for timing (s), variability (α), magnitude (β)

References

Pool, S., Vis, M., Seibert, J., 2018. Evaluating model performance: towards a non-parametric variant of the Kling-Gupta efficiency. Hydrological Sciences Journal 63, 1941–1953. https://doi.org/10.1080/02626667.2018.1552002

Streamfall.@bound — Macro

Bounds given metric between -1.0 and 1.0, where 1.0 is perfect fit.

Suitable for use with any metric that ranges from 1 to -∞.

References

Mathevet, T., Michel, C., Andréassian, V., Perrin, C., 2006. A bounded version of the Nash-Sutcliffe criterion for better model assessment on large sets of basins. IAHS-AISH Publication 307, 211–219. https://iahs.info/uploads/dms/13614.21–211-219-41-MATHEVET.pdf

Example

julia> import Streamfall: @bound, KGE
julia> @bound KGE([1,2], [3,2])
-0.35653767993482094

Streamfall.@mean_inverse — Macro

Applies mean inverse approach to a metric.

Suitable for use with any metric that ranges from 1 to -∞.

If using with other macros such as @normalize or @bound, these must come first.

References

Garcia, F., Folton, N., Oudin, L., 2017. Which objective function to calibrate rainfall–runoff models for low-flow index simulations? Hydrological Sciences Journal 62, 1149–1166. https://doi.org/10.1080/02626667.2017.1308511

Example

julia> import Streamfall: @normalize, @mean_inverse, KGE
julia> @normalize @mean_inverse KGE [1,2] [3,2] 1e-6
0.3193506006429825

Streamfall.@normalize — Macro

Normalizes given metric between 0.0 and 1.0, where 1.0 is perfect fit.

Suitable for use with any metric that ranges from 1 to -∞.

References

Nossent, J., Bauwens, W., 2012. Application of a normalized Nash-Sutcliffe efficiency to improve the accuracy of the Sobol’ sensitivity analysis of a hydrological model. EGU General Assembly Conference Abstracts 237.

Example

julia> import Streamfall: @normalize, KGE
julia> @normalize KGE([1,2], [3,2])
0.1111111111111111

Streamfall.@split — Macro

Applies split meta metric approach

If using with other macros such as @normalize or @bound, these must come first.

References

Fowler, K., Peel, M., Western, A., Zhang, L., 2018. Improved Rainfall-Runoff Calibration for Drying Climate: Choice of Objective Function. Water Resources Research 54, 3392–3408. https://doi.org/10.1029/2017WR022466

Example

julia> using Statistics
julia> import Streamfall: @normalize, @split, KGE
julia> @normalize @split KGE repeat([1,2], 365) repeat([3,2], 365) 365 mean
0.3217309561946589