Available metrics

Streamfall.ADJ_R2 — Method

Determine adjusted R^2

Arguments

obs::Vector : observations

sim::Vector : modeled results

p::Int : number of explanatory variables

Streamfall.BKGE — Method

Bounded KGE, bounded between -1 and 1.

Arguments

obs::Vector : observations

sim::Vector : modeled results

Streamfall.BmKGE — Method

Bounded modified KGE between -1 and 1.

Arguments

obs::Vector : observations

sim::Vector : modeled results

Streamfall.BnpKGE — Method

Bounded non-parametric KGE between -1 and 1.

Arguments

obs::Vector : observations

sim::Vector : modeled results

Streamfall.KGE — Method

Calculate the 2009 Kling-Gupta Efficiency (KGE) metric.

A KGE score of 1 means perfect fit. A score < -0.41 indicates that the mean of observations provides better estimates (see Knoben et al., 2019).

Note: Although similar, NSE and KGE cannot be directly compared.

Arguments

obs::Vector : observations

sim::Vector : modeled results

References

Gupta, H.V., Kling, H., Yilmaz, K.K., Martinez, G.F., 2009. Decomposition of the mean squared error and NSE performance criteria: Implications for improving hydrological modelling. Journal of Hydrology 377, 80–91. https://doi.org/10.1016/j.jhydrol.2009.08.003

Knoben, W.J.M., Freer, J.E., Woods, R.A., 2019. Technical note: Inherent benchmark or not? Comparing Nash-Sutcliffe and Kling-Gupta efficiency scores (preprint). Catchment hydrology/Modelling approaches. https://doi.org/10.5194/hess-2019-327

Streamfall.LME — Method

Liu Mean Efficiency metric.

Arguments

obs::Vector : observations

sim::Vector : modeled results

References

Liu, D., 2020. A rational performance criterion for hydrological model. Journal of Hydrology 590, 125488. https://doi.org/10.1016/j.jhydrol.2020.125488

Streamfall.NKGE — Method

Normalized KGE between 0 and 1.

Arguments

obs::Vector : observations

sim::Vector : modeled results

Streamfall.NNSE — Method

Normalized Nash-Sutcliffe Efficiency score (bounded between 0 and 1).

References

Nossent, J., Bauwens, W., 2012. Application of a normalized Nash-Sutcliffe efficiency to improve the accuracy of the Sobol’ sensitivity analysis of a hydrological model. EGU General Assembly Conference Abstracts 237.

Streamfall.NSE — Method

The Nash-Sutcliffe Efficiency score

Streamfall.NmKGE — Method

Normalized modified KGE between 0 and 1.

Arguments

obs::Vector : observations

sim::Vector : modeled results

Streamfall.NnpKGE — Method

Normalized non-parametric KGE between 0 and 1.

Arguments

obs::Vector : observations

sim::Vector : modeled results

Streamfall.R2 — Method

Coefficient of determination (R^2)

Aliases NSE()

Streamfall.RMSE — Method

Root Mean Square Error

Streamfall.mKGE — Method

Calculate the modified KGE metric (2012).

Also known as KGE prime (KGE').

Arguments

obs::Vector: observations

sim::Vector : modeled results

References

Kling, H., Fuchs, M., Paulin, M., 2012. Runoff conditions in the upper Danube basin under an ensemble of climate change scenarios. Journal of Hydrology 424–425, 264–277. https://doi.org/10.1016/j.jhydrol.2012.01.011

Streamfall.naive_split_metric — Method

Naive approach to split metrics.

Split metrics are a meta-objective optimization approach which segments data into subperiods. The objective function is calculated for each subperiod and then recombined. The approach addresses the lack of consideration of dry years with least-squares.

In Fowler et al., [1] the subperiod is one year. This method is "naive" in that the time series is partitioned into N chunks of n_members and does not consider date/time.

Arguments

obs::Vector : Historic observations to compare against

sim::Vector : Modeled time series

n_members::Int : number of members per chunk, defaults to 365

metric::Function : Objective function to apply, defaults to NNSE

comb_method::Function : Recombination method, defaults to mean

References

Fowler, K., Peel, M., Western, A., Zhang, L., 2018. Improved Rainfall-Runoff Calibration for Drying Climate: Choice of Objective Function. Water Resources Research 54, 3392–3408. https://doi.org/10.1029/2017WR022466

Streamfall.npKGE — Method

Calculate the non-parametric Kling-Gupta Efficiency (KGE) metric.

Arguments

obs::Vector : observations

sim::Vector : modeled results

References

Pool, S., Vis, M., Seibert, J., 2018. Evaluating model performance: towards a non-parametric variant of the Kling-Gupta efficiency. Hydrological Sciences Journal 63, 1941–1953. https://doi.org/10.1080/02626667.2018.1552002