Implemented node types

IHACRES

Streamfall.param_info — Method

param_info(node::IHACRESNode; with_level::Bool = true)::Tuple

Extract node parameter names, values, and bounds for IHACRESNode types.

Streamfall.reset! — Method

reset!(s_node::IHACRESNode)::Nothing

Reset node. Clears all states back to their initial values.

Streamfall.run_node! — Method

run_node!(s_node::BilinearNode,
          rain::Float64,
          evap::Float64,
          inflow::Float64,
          ext::Float64,
          gw_exchange::Float64,
          timestep::Union{Int, Nothing})::Tuple{Float64, Float64}

Run node for a given time step.

Streamfall.run_node! — Method

run_node!(s_node::BilinearNode,
          rain::Float64,
          evap::Float64,
          inflow::Float64,
          ext::Float64,
          gw_exchange::Float64=0.0;
          current_store::Union{Nothing, Float64}=nothing,
          quick_store::Union{Nothing, Float64}=nothing,
          slow_store::Union{Nothing, Float64}=nothing,
          gw_store::Union{Nothing, Float64}=nothing)::Tuple{Float64, Float64}

Run node with ET data to calculate outflow and update state.

Arguments

s_node::BilinearNode : IHACRESNode

rain : rainfall for time step

evap : evapotranspiration for time step

inflow : inflow from previous node

ext : irrigation and other water extractions

gw_exchange : flux in ML where positive is contribution to stream, negative is infiltration

current_store : replacement cmd state value (uses last known state if not provided)

quick_store : replacement quick store state value

slow_store : replacement slow store state value

gw_store : replacement groundwater store state value

Returns

float, outflow from node, stream level

Streamfall.run_node_with_temp! — Function

run_node_with_temp!(s_node::BilinearNode,
                    rain::Float64,
                    temp::Float64,
                    inflow::Float64,
                    ext::Float64,
                    gw_exchange::Float64=0.0;
                    current_store=nothing,
                    quick_store=nothing,
                    slow_store=nothing
                    gw_store=nothing)::Tuple{Float64, Float64}

Run node with temperature data to calculate outflow and update state.

Streamfall.update_params! — Method

update_params!(node::BilinearNode, d::Float64, d2::Float64, e::Float64, f::Float64,
               a::Float64, b::Float64, s_coef::Float64, alpha::Float64)::Nothing

Update model parameters.

Streamfall.update_params! — Method

update_params!(node::BilinearNode{Param}, d::Float64, d2::Float64, e::Float64, f::Float64,
               a::Float64, b::Float64, s_coef::Float64, alpha::Float64,
               p1::Float64, p2::Float64, p3::Float64, p4::Float64, p5::Float64, p6::Float64, p7::Float64, p8::Float64, CTF::Float64)::Nothing

Update all parameters.

Streamfall.BilinearNode — Method

BilinearNode(node_id::String, area::Float64, d::Float64, d2::Float64, e::Float64, f::Float64, 
            a::Float64, b::Float64, s_coef::Float64, alpha::Float64, 
            store::Float64, quick::Float64, slow::Float64, gw_store::Float64)

Create a IHACRES node that adopts the bilinear CMD module.

IHACRES - Expuh

Streamfall.run_node! — Function

run_node!(s_node::ExpuhNode, rain::Float64, evap::Float64, 
          inflow::Float64, ext::Float64, gw_exchange::Float64=0.0;
          current_store=nothing,
          quick_store=nothing,
          slow_store=nothing,
        )::Tuple

Run given IHACRES ExpuhNode for a time step based on last known state.

Dam Level

Streamfall.run_node! — Function

Calculate outflow for the dam node for a single time step.

Parameters

node : DamNode

rain : rainfall in mm

et : evapotranspiration data in mm

irrig_ext : irrigation extractions

extractions : extraction data in ML

gw_flux : groundwater interaction

Returns

outflow from dam

Streamfall.run_node! — Method

run_node!(node::DamNode, climate::Climate, ts::Int; 
          inflow=nothing, extraction=nothing, exchange=nothing)

Run a specific node for a specified time step.

Arguments

node::DamNode :

climate::Climate :

ts::Int : current time step

inflow::DataFrame : Time series of inflows from any upstream node.

extraction::DataFrame : Time series of water orders (expects column of _releases)

exchange::DataFrame : Time series of groundwater flux

Streamfall.update_params! — Method

update_params!(node::DamNode, storage_coef::Float64)::Nothing

Streamfall.update_volume — Method

update_volume(volume, node_inflow, gamma, rain, evap, area, extractions, discharge, max_store)::Float64

Update dam volume for timestep.

Arguments

volume : current water volume in ML

node_inflow : inflow from previous node in ML

gamma : groundwater exchange (positive is gain from gw flow, negative is loss to infiltration)

rain : rainfall input

evap : evaporation loss

infiltration : infiltration loss

area : dam surface area in square kilometers

extractions : water extraction from dam in ML

discharge : discharge from dam in ML

max_store : maximum dam storage in ML

Returns

volume of water stored in dam

Streamfall.DamNode — Method

DamNode(node_id::String, spec::Dict)

Create DamNode from a given specification.