Methods to run a network or node

Streamfall.align_time_frame — Method

align_time_frame(timeseries::T...)

Subset an arbitrary number of DataFrames to their shared period of time.

Returns subsetted copy of data in same order as input.

Example

julia> climate, streamflow = align_time_frame(climate, streamflow)

Streamfall.find_common_timeframe — Method

find_common_timeframe(timeseries::T...)

Find common time frame between time series.

Requires that all DataFrames have a "Date" column.

Streamfall.run_basin! — Method

run_basin!(sn::StreamfallNetwork, climate::Climate; extraction=nothing, exchange=nothing)

Run scenario for an entire catchment/basin.

Streamfall.run_node! — Method

run_node!(node::NetworkNode, climate;
          inflow=nothing, extraction=nothing, exchange=nothing)

Run a specific node, and only that node, for all time steps.

Arguments

node::NetworkNode : Any Streamfall NetworkNode

climate::Climate : Climate data

inflow::Union{DataFrame, Vector, Number} : Inflow to node

extraction::Union{DataFrame, Vector, Number} : Extractions from this subcatchment

exchange::Union{DataFrame, Vector, Number} : Groundwater flux

Streamfall.run_node! — Method

run_node!(sn::StreamfallNetwork, node_id::Int, climate::Climate, timestep::Int;
          extraction::Union{DataFrame, Nothing}=nothing,
          exchange::Union{DataFrame, Nothing}=nothing)

Run a model attached to a node for a given time step. Recurses upstream as needed.

Arguments

sn::StreamfallNetwork

node_id::Int

climate::Climate

timestep::Int

inflow::DataFrame : Additional inflow to consider (in ML/timestep)

extraction::DataFrame : Volume of water to be extracted (in ML/timestep)

exchange::DataFrame : Volume of flux (in ML/timestep), where negative values are losses to the groundwater system

Streamfall.run_node! — Method

run_node!(sn::StreamfallNetwork, node_id::Int, climate::Climate;
          extraction=nothing, exchange=nothing)::Nothing

Generic run method that runs a model attached to a given node for all time steps. Recurses upstream as needed.

Arguments

sn::StreamfallNetwork

node_id::Int : node to run in the network

climate::Climate : Climate object holding rainfall and evaporation data (or temperature)

extraction::DataFrame : water orders for each time step (defaults to nothing)

exchange::DataFrame : exchange with groundwater system at each time step (defaults to nothing)